Use EAGER loading of store team user
This makes doctrine to create an inner join on the user table thus
getting rid of entries that break the foreign key.

This should work for now (with the broken data) but should be considered
hacky & unstable as that doctrine behaviour might change and/or it is
easy to generate other cases where it might break again.